api organized / minor edits in views.py and in game and options templates / answers are now randomized also added slider for questions count
