--- api/Category.py
@@ -1,30 +1,0 @@
-MAIN_CATEGORIES = {  # Sõnastik, mis hoiab peamisi kategooriaid ja alamkategooriaid koos nende ID-dega
-    "General Knowledge": 9,
-    "Entertainment": {
-        "Books": 10,
-        "Film": 11,
-        "Music": 12,
-        # "Musicals & Theatres": 13# Eemaldatud kuna kokku küsimusi 32 igas raskusastmes vähe
-        "Television": 14,
-        "Video Games": 15,
-        "Board Games": 16,
-        "Comics": 29,
-        "Japanese Anime & Manga": 31,
-        "Cartoon & Animations": 32,
-    },
-    "Science": {
-        "Nature": 17,
-        "Computers": 18,
-        "Mathematics": 19, # E: 14 küsimust. Tiim otsustab
-        # "Gadgets": 30, Pole üldse sõnu kõiki küsimusi kokku alla 30
-    },
-    "Mythology": 20, # H: 13 küsimust. Tiim otsustab
-    "Sports": 21,
-    "Geography": 22,
-    "History": 23,
-    "Politics": 24,
-    # "Art": 25, Jälle 30 küsimuse ring. kõikides raskusastmetes vähe küsimusi
-    # "Celebrities": 26, M: üle 30 küsimuse, aga teised alla igasuguse arvestuse.
-    "Animals": 27,
-    "Vehicles": 28,
-}
