+++ templates/options.html
@@ -24,6 +24,18 @@
             {% endfor %}
         </select>
 
+        <label for="amount">Number of Questions:</label>
+        <input type="range" name="amount" id="amount" value="10" min="1" max="15">
+        <span id="amount-value">10</span>
+
+        <script>
+        const amountSlider = document.getElementById('amount');
+        const amountValue = document.getElementById('amount-value');
+
+        amountSlider.addEventListener('input', () => {
+        amountValue.textContent = amountSlider.value;
+        });
+        </script>
 
         <button type="submit">Start Game</button>
     </form>
